back-end/requirements.txt
alembic==1.5.8
click==7.1.2
Flask==1.1.2
Flask-Classful==0.14.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.0
flask-marshmallow==0.14.0
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.9
Werkzeug==1.0.1
flask_cors
pandas
webargs==7.0.1
flask_jwt_extended
sklearn

marshmallow~=3.12.1